test: 세션 id 없이 스터디룸 입장 테스트
